encryption added (deterministic variant, tink lib)
