+++ moneybox/api/views/auth.py
@@ -6,7 +6,7 @@ from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
 
-from api.crypto_temp import encrypt_token
+from api.encryption import encrypt_token
 from api.serializers import APIUserSerializer
 from moneybox.settings import AUTH_HEADER
 from users.models import APIUser
