+++ requirements/requirements.dev.txt
@@ -30,6 +30,7 @@ requests==2.28.2
 ruff==0.0.285
 six==1.16.0
 sqlparse==0.4.3
+tink==1.8.0
 typing_extensions==4.5.0
 tzdata==2023.3
 tzlocal==4.3
