+++ requirements/requirements.txt
@@ -28,6 +28,7 @@ redis==4.6.0
 requests==2.28.2
 six==1.16.0
 sqlparse==0.4.3
+tink==1.8.0
 typing_extensions==4.5.0
 tzdata==2023.3
 tzlocal==4.3
